Exception Handling

+++ src/main/java/com/leonardo/desafio03JSP/services/ClientService.java
@@ -3,7 +3,10 @@ package com.leonardo.desafio03JSP.services;
 import com.leonardo.desafio03JSP.DTO.ClientDTO;
 import com.leonardo.desafio03JSP.entities.Client;
 import com.leonardo.desafio03JSP.repositories.ClientRepository;
+import com.leonardo.desafio03JSP.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,7 +27,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        ClientDTO dto = new ClientDTO(repository.findById(id).get());
+        ClientDTO dto = new ClientDTO(repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Resource Not Found!")));
         return dto;
     }
 
@@ -38,14 +42,22 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyClientDtoToClient(entity, dto);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyClientDtoToClient(entity, dto);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource Not Found!");
+        }
+
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource Not Found!");
+        }
         repository.deleteById(id);
     }
